Fix destroy user photo and post photo bug

+++ app/Http/Controllers/AdminPostsController.php
@@ -143,7 +143,10 @@ class AdminPostsController extends Controller
 
         $post = Post::findOrFail($id);
 
-        unlink(public_path() . $post->photo->file);
+        if($post->photo) {
+            unlink(public_path() . $post->photo->file);
+        }
+
 
         $post->delete();
